Combine nativeVersionUpdater into bumpVersion
